Use relative path to module for archive files
This removes the assumption about Terraform making a .terraform directory in the current working directory, and removes the need to use md5 to handle multiple module versions in the same project as Terraform deals with that itself. It now assumes that the module is located within the current working directory.